rename: not found 폴더명 케밥케이스로 변경 (#35)
